Updated code formats and removed useless codes
