--- codes/typescript/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "module": "commonjs",
-    "strict": true,
-    "noEmit": false,
-    "sourceMap": true,
-    "baseUrl": ".",
-    "esModuleInterop": true,
-    "downlevelIteration": true,
-  },
-  "include": ["./**/*"],
-}
